+++ spec/lib/karafka/contracts/config_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
-RSpec.describe Karafka::Schemas::Config do
-  subject(:schema) { described_class.new }
+RSpec.describe Karafka::Contracts::Config do
+  subject(:contract) { described_class.new }
 
   let(:config) do
     {
@@ -13,20 +13,20 @@ RSpec.describe Karafka::Schemas::Config do
   end
 
   context 'when config is valid' do
-    it { expect(schema.call(config)).to be_success }
+    it { expect(contract.call(config)).to be_success }
   end
 
   context 'when we validate client_id' do
     context 'when client_id is nil' do
       before { config[:client_id] = nil }
 
-      it { expect(schema.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
     end
 
     context 'when client_id is not a string' do
       before { config[:client_id] = 2 }
 
-      it { expect(schema.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
     end
   end
 
@@ -34,19 +34,19 @@ RSpec.describe Karafka::Schemas::Config do
     context 'when shutdown_timeout is nil' do
       before { config[:shutdown_timeout] = nil }
 
-      it { expect(schema.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
     end
 
     context 'when shutdown_timeout is not an int' do
       before { config[:shutdown_timeout] = 2.1 }
 
-      it { expect(schema.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
     end
 
     context 'when shutdown_timeout is less then 0' do
       before { config[:shutdown_timeout] = -2 }
 
-      it { expect(schema.call(config)).not_to be_success }
+      it { expect(contract.call(config)).not_to be_success }
     end
   end
 end