+++ spec/lib/karafka/routing/builder_spec.rb
@@ -1,7 +1,7 @@
 # frozen_string_literal: true
 
 RSpec.describe Karafka::Routing::Builder do
-  subject(:builder) { described_class.instance }
+  subject(:builder) { described_class.new }
 
   ATTRIBUTES = %i[
     consumer