+++ spec/lib/karafka/server_spec.rb
@@ -3,58 +3,60 @@
 RSpec.describe Karafka::Server do
   subject(:server_class) { described_class }
 
+  let(:process) { Karafka::App.config.internal.process }
+
   describe '#run' do
     after { server_class.run }
 
     before do
       allow(Karafka::App).to receive(:run!)
-      allow(Karafka::Fetcher).to receive(:call)
+      allow(Karafka::App.config.internal.fetcher).to receive(:call)
     end
 
     context 'when we want to run in supervision' do
       it 'runs in supervision, start consuming' do
-        expect(Karafka::Process.instance).to receive(:on_sigint)
-        expect(Karafka::Process.instance).to receive(:on_sigquit)
-        expect(Karafka::Process.instance).to receive(:on_sigterm)
+        expect(process).to receive(:on_sigint)
+        expect(process).to receive(:on_sigquit)
+        expect(process).to receive(:on_sigterm)
       end
     end
 
     context 'when sigint is received' do
       before do
-        allow(Karafka::Process.instance).to receive(:supervise)
-        allow(Karafka::Process.instance).to receive(:on_sigquit)
-        allow(Karafka::Process.instance).to receive(:on_sigterm)
+        allow(process).to receive(:supervise)
+        allow(process).to receive(:on_sigquit)
+        allow(process).to receive(:on_sigterm)
       end
 
       it 'defines a proper action for sigint' do
         expect(described_class).to receive(:stop_supervised)
-        expect(Karafka::Process.instance).to receive(:on_sigint).and_yield
+        expect(process).to receive(:on_sigint).and_yield
       end
     end
 
     context 'when sigquit is received' do
       before do
-        allow(Karafka::Process.instance).to receive(:supervise)
-        allow(Karafka::Process.instance).to receive(:on_sigint)
-        allow(Karafka::Process.instance).to receive(:on_sigterm)
+        allow(process).to receive(:supervise)
+        allow(process).to receive(:on_sigint)
+        allow(process).to receive(:on_sigterm)
       end
 
       it 'defines a proper action for sigquit' do
         expect(described_class).to receive(:stop_supervised)
-        expect(Karafka::Process.instance).to receive(:on_sigquit).and_yield
+        expect(process).to receive(:on_sigquit).and_yield
       end
     end
 
     context 'when sigterm is received' do
       before do
-        allow(Karafka::Process.instance).to receive(:supervise)
-        allow(Karafka::Process.instance).to receive(:on_sigint)
-        allow(Karafka::Process.instance).to receive(:on_sigquit)
+        allow(process).to receive(:supervise)
+        allow(process).to receive(:on_sigint)
+        allow(process).to receive(:on_sigquit)
       end
 
       it 'defines a proper action for sigterm' do
         expect(described_class).to receive(:stop_supervised)
-        expect(Karafka::Process.instance).to receive(:on_sigterm).and_yield
+        expect(process).to receive(:on_sigterm).and_yield
       end
     end
   end
@@ -63,9 +65,9 @@ RSpec.describe Karafka::Server do
     after { server_class.send(:run_supervised) }
 
     it 'expect to supervise and run' do
-      expect(Karafka::Process.instance).to receive(:supervise)
+      expect(process).to receive(:supervise)
       expect(Karafka::App).to receive(:run!)
-      expect(Karafka::Fetcher).to receive(:call)
+      expect(Karafka::App.config.internal.fetcher).to receive(:call)
     end
   end